Changes to injection: Made code more general for the cases of multiple injection sources and multiple species. Now, the code is valid for the case where multiple injection source are close to each other and in the case where more than one species is injected from any single source. The change included adding the variables inj_ns and inj_ninj, which are either 1 (in the case of single species per source and non-touching sources) or top.ns and top.ninject. The arrays inj_prev and inj_np now have the shape (0:inj_nx,0:inj_ny,inj_ninj,inj_ns). When sources are touching, inj_prev and inj_np for each source are independent. With multiple species per source, again, each species has its own inj_prev and inj_np. The arrays inj_area and inj_rho now have dimension (0:inj_nx,0:inj_ny,inj_ninj), so the quantities are stored seperately for each source.
The code automatically checks for the exceptional caes, multiple species
per source, and touching sources, and sets inj_ns and inj_ninj accordingly.

In the case where there are multiple sources too close to each other, the
particle array pid is used to store the number of the injection source
from which the particles are born. This is done since resolving which
source a particle came from can not be done when the sources are too
close to each other. In other cases, in order to save memory, the pid
array is not used since the source number can be obtained from the
particle's position using the array inj_id.

One last change was removal of an 'if' test in the inj_sete3d routines
that was checking if a particle was within the radial extent of a
source.  This check was not correct in cases where the beam diverges
coming off of the source or converges coming off of an annular source.
For normal operation of the code, the check is not needed anyway.