Nunerous changes to the space-chare limited injection code:
Added additional flags to control the space-charge limited injection
 - linj_spherical: when true, use spherical correction in emitted current
 - linj_enormcl: when true, use E normal from the Child-Langmuir solution near
                 the emitter surface
 - linj_eperp: when true, apply perpendicular, or tangential, field to the
               particles near the emitter surface

Added vbeam_s to the initialization of vz.

Now, particles are given a small initial kick, i.e. an initial normal
velocity, equal to the velocity from the Child-Langmuir solution calculated
for each particle at a fractional time-step. The purpose of the kick is to get
the particles started when the Child-Langmuir solution for E normal is used -
in that case the E normal is zero at the emitter surface.

The pid array is now used always, storing the fractional time step that each
particle undergoes during its initialization (pid still also stores the number
of the injection source the particle was emitted from).

Much general code clean up and simplification was done.