This version has one functionality enhancement and two bug fixes:
(1) There is a new optional argument to the constructor for InjectionGaussLaw,
namely "relax".   If relax is set, the injection number at each time step is
set by relaxation from old values, rnn = relax*rnn + (1-relax)*rnn_old.
This option introduces some time smoothing of the injection, which in
some testing with the e.m. solver has been essential for numerical stability.

(2) The indices used to construct the fields on dual cells for construction
of the Gauss-Law charge for the e.m. solver has been generalized to
arbitrary numbers of guard cells (the number is taken from those specified
for the e.m. solver).  Before, the indexing was broken unless the number of
guard cells was everywhere 3.

(3) The correct metric factors are now used in RZ geometry for evaluating
Gaus's Law.