Cleaned up parallel exchange of data for MeshRefinement
Now, a dictionary which only contains the information needed is passed
to neighbors, instead of the whole class.
This was needed for the EM case since sending the whole class was
causing problems. This is a better way of doing it anyway since it
reduces the amount of data exchanged.