Fixed how fetchfield is handle before finalization
Now, the MR and EM field solvers return immediately if the solvers
have not been finalized when fetchfield is called. This is a rational
way of handling this since if the solver has not been finalized, then
there is no field to fetch. This avoids problems of trying to do the
finalize during the fetchfield. For instance, in parallel, if some of
the processors have no particles, fetchfield isn't called on those
processors, but is called on others, leading to a lock up since finalize
has global operations. This is what the ES solvers already do, returning
immediately if the field arrays have not been setup.